sync spec infra with caseflow
