+++ spec/rails_helper.rb
@@ -24,92 +24,6 @@ require_relative "../app/exceptions/standard_error"
 #
 Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
 
-require "selenium-webdriver"
-require "capybara"
-
-Sniffybara::Driver.path_exclusions << /samlva/
-Sniffybara::Driver.configuration_file = File.expand_path("../support/VA-axe-configuration.json", __FILE__)
-Sniffybara::Driver.issue_id_exceptions += []
-
-download_directory = Rails.root.join("tmp", "downloads_all")
-cache_directory = Rails.root.join("tmp", "browser_cache_all")
-
-FileUtils.mkpath download_directory unless File.directory?(download_directory)
-if File.directory?(cache_directory)
-  FileUtils.rm_r cache_directory
-else
-  Dir.mkdir cache_directory
-end
-
-Capybara.register_driver(:virtual_framebuffer_chrome) do |app|
-  chrome_options = ::Selenium::WebDriver::Chrome::Options.new
-
-  chrome_options.add_preference(:download,
-                                prompt_for_download: false,
-                                default_directory: download_directory)
-
-  chrome_options.add_preference(:browser,
-                                disk_cache_dir: cache_directory)
-
-  options = {
-    port: 51_674,
-    browser: :chrome,
-    options: chrome_options
-  }
-
-  Sniffybara::Driver.current_driver = Sniffybara::Driver.new(app, options)
-end
-Capybara.default_driver = :virtual_framebuffer_chrome
-
-# Convenience methods for stubbing current user
-module StubbableUser
-  module ClassMethods
-    def stub=(user)
-      @stub = user
-    end
-
-    def authenticate!(options = {})
-      css_id = options[:css_id] || "123123"
-      user_name = options[:user_name] || "first last"
-      Functions.grant!("System Admin", users: [css_id]) if options[:roles]&.include?("System Admin")
-
-      self.stub = find_or_create_by(css_id: css_id, station_id: "116").tap do |u|
-        u.name = user_name
-        u.email = "test@gmail.com"
-        u.roles = options[:roles] || ["Download eFolder"]
-        u.save
-        RequestStore.store[:current_user] = u
-      end
-    end
-
-    def tester!(options = {})
-      self.stub = find_or_create_by(css_id: ENV["TEST_USER_ID"], station_id: "116").tap do |u|
-        u.name = "first last"
-        u.email = "test@gmail.com"
-        u.roles = options[:roles] || ["Download eFolder"]
-        u.save
-      end
-    end
-
-    def unauthenticate!
-      Functions.delete_all_keys!
-      RequestStore.store[:current_user] = nil
-      self.stub = nil
-    end
-
-    def from_session_and_request(session, request)
-      @stub || super(session, request)
-    end
-  end
-
-  def self.prepended(base)
-    class << base
-      prepend ClassMethods
-    end
-  end
-end
-User.prepend(StubbableUser)
-
 module RandomHelper
   def self.valid_document_id
     "{#{SecureRandom.uuid.upcase}}"
