+++ spec/support/VA-axe-run-configuration.json
@@ -1,0 +1,48 @@
+{
+  "runOnly": {
+    "type": "tag",
+    "values": [
+      "section508",
+      "wcag2a",
+      "wcag2aa",
+    ],
+  },
+  "rules": {
+    "aria-roles": {
+      "enabled": false,
+    },
+    "aria-required-parent": {
+      "enabled": false,
+    },
+    "aria-allowed-attr": {
+      "enabled": false,
+    },
+    "aria-input-field-name": {
+      "enabled": false,
+    },
+    "aria-required-children": {
+      "enabled": false,
+    },
+    "list": {
+      "enabled": false,
+    },
+    "definition-list": {
+      "enabled": false,
+    },
+    "scrollable-region-focusable": {
+      "enabled": false,
+    },
+    "listitem": {
+      "enabled": false,
+    },
+    "duplicate-id-aria": {
+      "enabled": false,
+    },
+    "color-contrast": {
+      "enabled": false,
+    },
+    "duplicate-id-active": {
+      "enabled": false,
+    }
+  },
+}
